requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.12
Django==5.1.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
sqlparse==0.5.3
tzdata==2025.1
